+++ MaxMind.MinFraud.UnitTest/Request/TestHelper.cs
@@ -113,7 +113,8 @@ namespace MaxMind.MinFraud.UnitTest.Request
                     transactionId: "txn3134133",
                     shopId: "s2123",
                     time: new DateTimeOffset(2014, 4, 12, 23, 20, 50, 52, new TimeSpan(0)),
-                    type: EventType.Purchase
+                    type: EventType.Purchase,
+                    party: null
                 ),
                 order:
                 new Order(